Solve the issue with tangent space vectors
o dpdu and dpdv vectors in the structure DifferentialGeometry
treat as tangent and bitangent vectors in the kernels.
Actually there is a difference: dpdu/dpdv are need to be
oriented along increasing of uv coordinates, may be not
orthogonal and their length may be not equal to one.

The solution is:
1. Compute dpdu/dpdv in the form, as it need to be used in
screen space uv derivative calculation;
2. Calculate screen space uv derivative;
3. Replace dpdu/dpdv values to the tangent/bitangent values.